Move typscript test copy script to scripts folder and run once
