+++ packages/babel-generator/test/fixtures/typescript/class-abstract/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
