+++ packages/babel-generator/test/fixtures/typescript/class-declare/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
