+++ packages/babel-generator/test/fixtures/typescript/class-members-with-modifier-names/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
