+++ packages/babel-generator/test/fixtures/typescript/class-method-no-body/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
