+++ packages/babel-generator/test/fixtures/typescript/class-method-return-type/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
