+++ packages/babel-generator/test/fixtures/typescript/class-method-with-newline-without-body/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
