+++ packages/babel-generator/test/fixtures/typescript/class-modifiers-methods-async/options.json
@@ -1,0 +1,4 @@
+{
+  "sourceType": "module",
+  "plugins": ["typescript", "classProperties"]
+}
