+++ packages/babel-generator/test/fixtures/typescript/options.json
@@ -1,4 +1,4 @@
 {
   "sourceType": "module",
-  "plugins": ["typescript", "classProperties"]
+  "plugins": ["typescript"]
 }
