+++ packages/babel-generator/test/fixtures/typescript/regression-is-default-export/expected.js
@@ -1,0 +1,3 @@
+export type T = number;
+export enum E {}
+export interface I {}
