+++ packages/babel-generator/test/fixtures/typescript/regression-is-default-export/options.json
@@ -1,0 +1,3 @@
+{
+    "plugins": ["exportExtensions", "typescript"]
+}
