+++ packages/babel-generator/test/fixtures/typescript/type-arguments-whitespace/expected.js
@@ -1,0 +1,1 @@
+function f<T>() {}
